+++ tests/test_sco_client.py
@@ -13,6 +13,39 @@ class TestSCOClient(unittest.TestCase):
         """Initialize the SCO Client. Assumes that defautl Web API is active."""
         self.sco = scocli.SCOClient(api_url=API_URL)
 
+    def test_create_experiment_with_functional_data(self):
+        """Create experiment and assign functional data."""
+        # Create subject
+        subject = self.sco.subjects_create(
+            os.path.join(DATA_DIR, 'subjects/ernie.tar.gz'),
+            properties={'name' : 'Ernie'}
+        )
+        # Create image group
+        image_group = self.sco.image_groups_create(
+            os.path.join(DATA_DIR, 'images/images.tar.gz'),
+            options={
+                'pixels_per_degree' : 1,
+                'aperture_edge_width' : 0.9
+            },
+            properties={'name' : 'AutoUploadImages'}
+        )
+        # Create experiment
+        experiment = self.sco.experiments_create(
+            'Name',
+            subject.identifier,
+            image_group.identifier
+        )
+        # Upload fMRI file for experiment
+        self.sco.experiments_fmri_create(
+            experiment.url,
+            os.path.join(DATA_DIR, 'fmris/fmri.tar')
+        )
+        # Get the experiment and ansure that the associated fmri data has two
+        # files
+        experiment = self.sco.experiments_get(experiment.url)
+        fmri = experiment.fmri_data
+        self.assertEquals(len(fmri.data_files), 2)
+
     def test_create_experiment_with_missing_references(self):
         """Create experiment with incorrect subject and image group references.
         Ensure that exceptions are thrown as expected."""